Use assigned object as TerraformObject instead of copying properties to a new object
Enhance assigned object by setting its prototype rather than creating a new object and copying properties to it. This simplifies the TerraformObject implementation and more importantly means you can assign a local variable to reference the object in the same statement where you assign the object to a property of TerraformClass. This allows a compact syntax as demonstrated in the configuration spec